+++ two_d_game_ai/entities/generic_entities.py
@@ -6,6 +6,7 @@ from abc import ABC, abstractmethod
 from dataclasses import InitVar, dataclass, field
 from typing import TYPE_CHECKING
 
+from loguru import logger
 from pygame import Vector2
 
 if TYPE_CHECKING:
@@ -31,6 +32,8 @@ class GenericEntity(ABC):
 
     def __post_init__(self, position_from_sequence: Sequence[float]) -> None:
         self.position = Vector2(position_from_sequence)
+        if self.world and not self.is_inside_world_bounds:
+            logger.warning(f"{self!s}: outside World bounds.")
 
     def __hash__(self) -> int:
         if self.id is None:
@@ -41,6 +44,14 @@ class GenericEntity(ABC):
     def __str__(self) -> str:
         """Human-readable description."""
         return f"{type(self).__name__} '{self.name}', id={self.id}"
+
+    @property
+    def is_inside_world_bounds(self) -> bool:
+        """Return `True` if entity is inside the World bounds, else `False`."""
+        if self.world is None:
+            err_msg = f"Can't check {self!s} inside bounds. Add to World first."
+            raise ValueError(err_msg)
+        return self.world.location_is_inside_world_bounds(self.position)
 
     @abstractmethod
     def add_to_grid(self) -> None: