+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Historic and pre-release versions aren't necessarily included.
 
 - Bot has radius (not yet used for collision etc)
 - Obstacles (previously MovementBlocks) can be rectangles
+- Warn if entity position initially outside World 
 - View: improved colours and text rendering
 - Simplified and improved log messages
 - Build: PEP621-compatible `pyproject.toml`